+++ src/cpp/surface-generation/biomes/cubic-spline.hpp
@@ -99,8 +99,8 @@ public:
 
         float apply(C value) override {
             float loc = this->coordinate(value);
-            int32_t index = Mth::binarySearch(0, this->locations.size(), loc, this->locations) - 1;
-            int32_t lastIndex = this->locations.size() - 1;
+            int32_t index = Mth::binarySearch(0, (int32_t)this->locations.size(), loc, this->locations) - 1;
+            int32_t lastIndex = (int32_t)this->locations.size() - 1;
             if (index < 0) {
                 return this->values.at(0)->apply(value) + this->derivatives.at(0) * (loc - this->locations.at(0));
             } else if (index == lastIndex) {
